requirements.txt
Flask==1.1.4
Jinja2==2.11.3
psycopg2-binary==2.9
Werkzeug==1.0.1
boto3==1.17.101
openpyxl==3.0.7
markupsafe==2.0.1





